chore: extract components that display attribution info into reusable component
